dodat pregled rezervisanih lekova i appointment-a

+++ src/main/java/com/mrsisa/mrsisaprojekat/controller/PatientController.java
@@ -1,7 +1,9 @@
 package com.mrsisa.mrsisaprojekat.controller;
 
 import java.net.URI;
-import java.util.Collection;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.*;
 
 import com.mrsisa.mrsisaprojekat.dto.AppointmentDTO;
 import com.mrsisa.mrsisaprojekat.dto.PatientDTO;
@@ -27,7 +29,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Collection;
-import java.util.Set;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
@@ -57,7 +58,49 @@ public class PatientController {
 	@Autowired
 	private PharmacistService pharmacistService;
 	
-	
+
+	@GetMapping(value="/reservedMedication/{id}")
+	public ResponseEntity<Collection<PrescriptionMedicamentDTO>> getReservedMedication(@PathVariable("id") String id) {
+		Patient patient = patientService.getOneWithReservedMeds(id);
+
+		if(patient == null)
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+		List<PrescriptionMedicamentDTO> returns = new ArrayList<>();
+		for(PrescriptionMedicament pm : patient.getReservedMedicaments()) {
+			if(!pm.isDeleted() && !pm.isPurchased()) {
+				PrescriptionMedicamentDTO medicament = new PrescriptionMedicamentDTO(pm);
+				returns.add(medicament);
+			}
+
+		}
+
+		return new ResponseEntity<>(returns,HttpStatus.OK);
+
+
+	}
+
+
+
+	@GetMapping(value="/reservedAppointments/{id}")
+	public ResponseEntity<Collection<AppointmentDTO>> getReservedAppointments(@PathVariable("id") String id) {
+		Patient patient = patientService.getOneWithAppointments(id);
+
+		if (patient == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+
+		List<AppointmentDTO> returns = new ArrayList<>();
+		for(Appointment a : patient.getAppointments()) {
+			if(!a.isDone()) {
+				AppointmentDTO d = new AppointmentDTO(a);
+				returns.add(d);
+			}
+
+		}
+
+		return new ResponseEntity<>(returns,HttpStatus.OK);
+	}
 	
 	@PostMapping(consumes = "application/json")
 	@PreAuthorize("hasAnyRole('DERMATOLOGIST', 'PHARMACIST')")
@@ -173,17 +216,19 @@ public class PatientController {
 
 	}
 
-	@DeleteMapping(value = "/{id}")
+	@DeleteMapping(value = "cancelReservation/{id}")
 	@PreAuthorize("hasAnyRole('DERMATOLOGIST', 'PHARMACIST', 'PATIENT')")
 	public ResponseEntity<Long> cancelMedicamentReservation(@PathVariable("id") Long id) {
 		PrescriptionMedicament medicament = prescriptionMedicamentService.findOne(id);
 
 		if(medicament != null) {
-			prescriptionMedicamentService.delete(id);
-			return new ResponseEntity<>(id, HttpStatus.OK);
-		} else {
-			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+			if(LocalDate.now().plusDays(1).isBefore(medicament.getExpiryDate())) {
+				prescriptionMedicamentService.delete(id);
+				return new ResponseEntity<>(id, HttpStatus.OK);
+			}
 		}
+		return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+
 	}
 
 //	@PostMapping(path = "/reserve_appointment", consumes = "application/json")
@@ -229,13 +274,16 @@ public class PatientController {
 	//@PreAuthorize("hasAnyRole('DERMATOLOGIST', 'PHARMACIST')")
 	public ResponseEntity<Long> cancelExamination(@PathVariable("id") Long id) {
 		Appointment a = appointmentService.findOne(id);
+		LocalDateTime now = LocalDateTime.now().plusDays(1);
 
 		if(a != null) {
-			appointmentService.delete(a);
-			return new ResponseEntity<>(id, HttpStatus.OK);
-		} else {
-			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+			if(now.isBefore(a.getDate().atTime(a.getTermFrom()))) {
+				appointmentService.delete(a);
+				return new ResponseEntity<>(id, HttpStatus.OK);
+			}
 		}
+		return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+
 	}
 
 //	@GetMapping(value = "/{id}/appointments")
